server: remove old server
