+++ Cargo.toml
@@ -46,7 +46,6 @@ tracing-subscriber = { version = "0.3.18", features = ["env-filter"] }
 url = "2.5.2"
 urlencoding = "2.1.3"
 uuid = { version = "1.10.0", features = ["v4"] }
-warp = "0.3.7"
 xml5ever = "0.18.1"
 
 [dependencies.reqwest]
